requirements.txt
Flask
psycopg2-binary
gunicorn
Flask-WTF
Jinja2
Werkzeug 